Move trait into existing module
